+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Changelog
 ## 
+- 2025-05-24 17:16: Created Home.jsx with Header, Footer and ScrollToTop components. Created placeholders for other attractions
 - 2025-05-24 16:12: Initial commit